+++ src/main/java/cart/application/ProductService.java
@@ -1,13 +1,12 @@
 package cart.application;
 
-import cart.domain.Product;
 import cart.dao.ProductDao;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
-import org.springframework.stereotype.Service;
-
+import cart.domain.Product;
+import cart.dto.request.ProductRequest;
+import cart.dto.response.ProductResponse;
 import java.util.List;
 import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
 
 @Service
 public class ProductService {
@@ -19,8 +18,9 @@ public class ProductService {
     }
 
     public List<ProductResponse> getAllProducts() {
-        List<Product> products = productDao.getAllProducts();
-        return products.stream().map(ProductResponse::of).collect(Collectors.toList());
+        return productDao.getAllProducts().stream()
+                .map(ProductResponse::of)
+                .collect(Collectors.toList());
     }
 
     public ProductResponse getProductById(Long productId) {
@@ -29,12 +29,14 @@ public class ProductService {
     }
 
     public Long createProduct(ProductRequest productRequest) {
-        Product product = new Product(productRequest.getName(), productRequest.getPrice(), productRequest.getImageUrl());
+        Product product = new Product(productRequest.getName(), productRequest.getPrice(),
+                productRequest.getImageUrl());
         return productDao.createProduct(product);
     }
 
     public void updateProduct(Long productId, ProductRequest productRequest) {
-        Product product = new Product(productRequest.getName(), productRequest.getPrice(), productRequest.getImageUrl());
+        Product product = new Product(productRequest.getName(), productRequest.getPrice(),
+                productRequest.getImageUrl());
         productDao.updateProduct(productId, product);
     }
