--- src/main/java/cart/dto/CartItemRequest.java
@@ -1,16 +1,0 @@
-package cart.dto;
-
-public class CartItemRequest {
-    private Long productId;
-
-    public CartItemRequest() {
-    }
-
-    public CartItemRequest(Long productId) {
-        this.productId = productId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-}
